requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0

# Google API integrations
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gspread>=5.10.0

# Web interface
streamlit>=1.30.0
